add docs

+++ python_stt/automatic_sr.py
@@ -3,16 +3,24 @@ import os
 import speech_recognition as sr
 import ffmpeg
 
-def convert_stt(filename):
+def convert_stt(input_filename):
+
+    '''
+    input_filename: string representing the path to an audio recording file
+    '''
+
     recognizer = sr.Recognizer()
 
-    output_filename = '.'.join(filename.split('.')[:-1]) + '.wav'
+    # target filename specifying the destination audio codec
+    output_filename = '.'.join(input_filename.split('.')[:-1]) + '.wav'
 
-    stream = ffmpeg.input(filename)
-    stream = ffmpeg.output(stream, output_filename)
-    ffmpeg.run(stream)
+    # convert from original to destination codec if applicable
+    if input_filename != output_filename:
+        stream = ffmpeg.input(input_filename)
+        stream = ffmpeg.output(stream, output_filename)
+        ffmpeg.run(stream)
 
-    print("Successfully converted '{}' to '{}'.".format(filename, output_filename))
+        print("Successfully converted '{}' to '{}'.".format(input_filename, output_filename))
 
     with sr.AudioFile(output_filename) as source:
 
@@ -23,8 +31,12 @@ def convert_stt(filename):
             transcription = recognizer.recognize_google(audio_text)
             print('Transcription: "{}"'.format(transcription))
 
-            os.remove(filename)
-            os.remove(output_filename)
+            # delete original and converted audio files
+            if os.path.exists(input_filename):
+                os.remove(input_filename)
+            
+            if os.path.exists(output_filename):
+                os.remove(output_filename)
 
             return transcription
 
@@ -37,3 +49,4 @@ def convert_stt(filename):
 if __name__ == "__main__":
 
     test = convert_stt("uploads/test.ogg")
+    print(test)
